backend/requirements.txt
django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.8
django-celery-beat==2.6.0
python-dotenv==1.0.1
drf-yasg==1.21.7
django-cors-headers==4.7.0